Refractor the codes
