+++ rental_manager.rb
@@ -1,0 +1,11 @@
+require './rental'
+
+class RentalManager
+  def initialize
+
+  end
+  
+  def add_rental(book, person, date)
+    Rental.new(book, person, date)
+  end
+end
